refactor: split dsp.processing module in multiple modules
