+++ pyoae/cdpoae.py
@@ -33,7 +33,7 @@ from pyoae import helpers
 from pyoae.calib import MicroTransferFunction, OutputCalibration
 from pyoae.device.device_config import DeviceConfig
 from pyoae.dsp.containers import DpoaeMsrmtData
-from pyoae.dsp.processing import ContDpoaeProcessor
+from pyoae.dsp.continuous_dpoae import ContDpoaeProcessor
 from pyoae.generator import ContDpoaeStimulus
 from pyoae.msrmt_context import DpoaeMsrmtContext
 from pyoae.protocols import DpoaeMsrmtParams
